third_party/blink/web_tests/fast/xpath/4XPath/Borrowed/rs_20010831-expected.txt

PASS nodeset.snapshotLength is 2
PASS nodeset.snapshotItem(0).nodeValue is "normal text"
PASS nodeset.snapshotItem(1).nodeValue is "<cdatatext>"
PASS nodeset.singleNodeValue.nodeValue is "<cdatatext>"
PASS successfullyParsed is true

TEST COMPLETE
